+++ CHANGELOG.md
@@ -46,6 +46,9 @@
 * Benthic Transform Random Item:
   * Added config to change how many items are transformed per effect activation
 
+* Disable Random Skill:
+  * Added config to exclude specific skill slots from the effect
+
 **1.10.1 Changes:**
 
 <details>